requirements.txt
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.11.8
chardet==3.0.4
idna==2.10
mysql-connector==2.2.9
pandas==1.1.4
Pillow==8.0.1
python-dateutil==2.8.1
pytz==2020.4
requests==2.25.0
six==1.15.0
soupsieve==2.0.1
wxPython==4.1.1
